Extract several large pieces of code from resolver
