+++ Source/DafnyCore/Resolver/Resolver.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Dafny {
   public partial class Resolver {
     readonly BuiltIns builtIns;
 
-    ErrorReporter reporter;
+    public ErrorReporter reporter;
     ModuleSignature moduleInfo = null;
 
     public ErrorReporter Reporter => reporter;
@@ -3306,16 +3306,7 @@ namespace Microsoft.Dafny {
         }
         // Check that functions claiming to be abstemious really are, and check that 'older' parameters are used only when allowed
         foreach (var fn in ModuleDefinition.AllFunctions(declarations)) {
-          if (fn.Body != null) {
-            var abstemious = true;
-            if (Attributes.ContainsBool(fn.Attributes, "abstemious", ref abstemious) && abstemious) {
-              if (CoCallResolution.GuaranteedCoCtors(fn) == 0) {
-                reporter.Error(MessageSource.Resolver, fn, "the value returned by an abstemious function must come from invoking a co-constructor");
-              } else {
-                CheckDestructsAreAbstemiousCompliant(fn.Body);
-              }
-            }
-          }
+          new Abstemious(reporter).Check(fn);
           CheckOlderParameters(fn);
         }
         // Check that all == and != operators in non-ghost contexts are applied to equality-supporting types.
@@ -3604,7 +3595,7 @@ namespace Microsoft.Dafny {
             if (m.Body != null) {
               ComputeGhostInterest(m.Body, m.IsGhost, m.IsLemmaLike ? "a " + m.WhatKind : null, m);
               CheckExpression(m.Body, this, m);
-              DetermineTailRecursion(m);
+              new TailRecursion(reporter).DetermineTailRecursion(m);
             }
           } else if (member is Function) {
             var f = (Function)member;
@@ -3614,7 +3605,7 @@ namespace Microsoft.Dafny {
                 ExpressionTester.CheckIsCompilable(this, f.Body, f);
               }
               if (f.Body != null) {
-                DetermineTailRecursion(f);
+                new TailRecursion(reporter).DetermineTailRecursion(f);
               }
             } else {
               var m = f.ByMethodDecl;
@@ -3622,7 +3613,7 @@ namespace Microsoft.Dafny {
                 Contract.Assert(!m.IsGhost);
                 ComputeGhostInterest(m.Body, false, null, m);
                 CheckExpression(m.Body, this, m);
-                DetermineTailRecursion(m);
+                new TailRecursion(reporter).DetermineTailRecursion(m);
               } else {
                 // m should not be null, unless an error has been reported
                 // (e.g. function-by-method and method with the same name) 
@@ -3651,47 +3642,6 @@ namespace Microsoft.Dafny {
       }
     }
 
-    private void CheckDestructsAreAbstemiousCompliant(Expression expr) {
-      Contract.Assert(expr != null);
-      expr = expr.Resolved;
-      if (expr is MemberSelectExpr) {
-        var e = (MemberSelectExpr)expr;
-        if (e.Member.EnclosingClass is CoDatatypeDecl) {
-          var ide = Expression.StripParens(e.Obj).Resolved as IdentifierExpr;
-          if (ide != null && ide.Var is Formal) {
-            // cool
-          } else {
-            reporter.Error(MessageSource.Resolver, expr, "an abstemious function is allowed to invoke a codatatype destructor only on its parameters");
-          }
-          return;
-        }
-      } else if (expr is MatchExpr) {
-        var e = (MatchExpr)expr;
-        if (e.Source.Type.IsCoDatatype) {
-          var ide = Expression.StripParens(e.Source).Resolved as IdentifierExpr;
-          if (ide != null && ide.Var is Formal) {
-            // cool; fall through to check match branches
-          } else {
-            reporter.Error(MessageSource.Resolver, e.Source, "an abstemious function is allowed to codatatype-match only on its parameters");
-            return;
-          }
-        }
-      } else if (expr is BinaryExpr) {
-        var e = (BinaryExpr)expr;
-        if (e.ResolvedOp == BinaryExpr.ResolvedOpcode.EqCommon || e.ResolvedOp == BinaryExpr.ResolvedOpcode.NeqCommon) {
-          if (e.E0.Type.IsCoDatatype) {
-            reporter.Error(MessageSource.Resolver, expr, "an abstemious function is not only allowed to check codatatype equality");
-            return;
-          }
-        }
-      } else if (expr is StmtExpr) {
-        var e = (StmtExpr)expr;
-        // ignore the statement part
-        CheckDestructsAreAbstemiousCompliant(e.E);
-        return;
-      }
-      expr.SubExpressions.Iter(CheckDestructsAreAbstemiousCompliant);
-    }
 
     /// <summary>
     /// Add edges to the callgraph.
@@ -6569,485 +6519,6 @@ namespace Microsoft.Dafny {
     }
     #endregion Visitors
 
-    // ------------------------------------------------------------------------------------------------------
-    // ----- CheckTypeInference -----------------------------------------------------------------------------
-    // ------------------------------------------------------------------------------------------------------
-    // The CheckTypeInference machinery visits every type in a given part of the AST (for example,
-    // CheckTypeInference(Expression) does so for an Expression and CheckTypeInference_Member(MemberDecl)
-    // does so for a MemberDecl) to make sure that all parts of types were fully inferred.
-    #region CheckTypeInference
-    private void CheckTypeInference_Member(MemberDecl member) {
-      if (member is ConstantField) {
-        var field = (ConstantField)member;
-        if (field.Rhs != null) {
-          CheckTypeInference(field.Rhs, new NoContext(member.EnclosingClass.EnclosingModuleDefinition));
-        }
-        CheckTypeInference(field.Type, new NoContext(member.EnclosingClass.EnclosingModuleDefinition), field.tok, "const");
-      } else if (member is Method) {
-        var m = (Method)member;
-        foreach (var formal in m.Ins) {
-          if (formal.DefaultValue != null) {
-            CheckTypeInference(formal.DefaultValue, m);
-          }
-        }
-        m.Req.Iter(mfe => CheckTypeInference_MaybeFreeExpression(mfe, m));
-        m.Ens.Iter(mfe => CheckTypeInference_MaybeFreeExpression(mfe, m));
-        CheckTypeInference_Specification_FrameExpr(m.Mod, m);
-        CheckTypeInference_Specification_Expr(m.Decreases, m);
-        if (m.Body != null) {
-          CheckTypeInference(m.Body, m);
-        }
-      } else if (member is Function) {
-        var f = (Function)member;
-        foreach (var formal in f.Formals) {
-          if (formal.DefaultValue != null) {
-            CheckTypeInference(formal.DefaultValue, f);
-          }
-        }
-        var errorCount = reporter.Count(ErrorLevel.Error);
-        f.Req.Iter(e => CheckTypeInference(e.E, f));
-        f.Ens.Iter(e => CheckTypeInference(e.E, f));
-        f.Reads.Iter(fe => CheckTypeInference(fe.E, f));
-        CheckTypeInference_Specification_Expr(f.Decreases, f);
-        if (f.Body != null) {
-          CheckTypeInference(f.Body, f);
-        }
-        if (errorCount == reporter.Count(ErrorLevel.Error)) {
-          if (f is ExtremePredicate cop) {
-            CheckTypeInference_Member(cop.PrefixPredicate);
-          } else if (f.ByMethodDecl != null) {
-            CheckTypeInference_Member(f.ByMethodDecl);
-          }
-        }
-      }
-    }
-
-    private void CheckTypeInference_MaybeFreeExpression(AttributedExpression mfe, ICodeContext codeContext) {
-      Contract.Requires(mfe != null);
-      Contract.Requires(codeContext != null);
-      foreach (var e in Attributes.SubExpressions(mfe.Attributes)) {
-        CheckTypeInference(e, codeContext);
-      }
-      CheckTypeInference(mfe.E, codeContext);
-    }
-    private void CheckTypeInference_Specification_Expr(Specification<Expression> spec, ICodeContext codeContext) {
-      Contract.Requires(spec != null);
-      Contract.Requires(codeContext != null);
-      foreach (var e in Attributes.SubExpressions(spec.Attributes)) {
-        CheckTypeInference(e, codeContext);
-      }
-      spec.Expressions.Iter(e => CheckTypeInference(e, codeContext));
-    }
-    private void CheckTypeInference_Specification_FrameExpr(Specification<FrameExpression> spec, ICodeContext codeContext) {
-      Contract.Requires(spec != null);
-      Contract.Requires(codeContext != null);
-      foreach (var e in Attributes.SubExpressions(spec.Attributes)) {
-        CheckTypeInference(e, codeContext);
-      }
-      spec.Expressions.Iter(fe => CheckTypeInference(fe.E, codeContext));
-    }
-    void CheckTypeInference(Expression expr, ICodeContext codeContext) {
-      Contract.Requires(expr != null);
-      Contract.Requires(codeContext != null);
-      PartiallySolveTypeConstraints(true);
-      var c = new CheckTypeInference_Visitor(this, codeContext);
-      c.Visit(expr);
-    }
-    void CheckTypeInference(Type type, ICodeContext codeContext, IToken tok, string what) {
-      Contract.Requires(type != null);
-      Contract.Requires(codeContext != null);
-      Contract.Requires(tok != null);
-      Contract.Requires(what != null);
-      PartiallySolveTypeConstraints(true);
-      var c = new CheckTypeInference_Visitor(this, codeContext);
-      c.CheckTypeIsDetermined(tok, type, what);
-    }
-    void CheckTypeInference(Statement stmt, ICodeContext codeContext) {
-      Contract.Requires(stmt != null);
-      Contract.Requires(codeContext != null);
-      PartiallySolveTypeConstraints(true);
-      var c = new CheckTypeInference_Visitor(this, codeContext);
-      c.Visit(stmt);
-    }
-    class CheckTypeInference_Visitor : ResolverBottomUpVisitor {
-      readonly ICodeContext codeContext;
-      public CheckTypeInference_Visitor(Resolver resolver, ICodeContext codeContext)
-        : base(resolver) {
-        Contract.Requires(resolver != null);
-        Contract.Requires(codeContext != null);
-        this.codeContext = codeContext;
-      }
-
-      protected override void VisitOneStmt(Statement stmt) {
-        if (stmt is VarDeclStmt) {
-          var s = (VarDeclStmt)stmt;
-          foreach (var local in s.Locals) {
-            CheckTypeIsDetermined(local.Tok, local.Type, "local variable");
-            CheckTypeArgsContainNoOrdinal(local.Tok, local.type);
-          }
-        } else if (stmt is VarDeclPattern) {
-          var s = (VarDeclPattern)stmt;
-          s.LocalVars.Iter(local => CheckTypeIsDetermined(local.Tok, local.Type, "local variable"));
-          s.LocalVars.Iter(local => CheckTypeArgsContainNoOrdinal(local.Tok, local.Type));
-
-        } else if (stmt is ForallStmt) {
-          var s = (ForallStmt)stmt;
-          s.BoundVars.Iter(bv => CheckTypeIsDetermined(bv.tok, bv.Type, "bound variable"));
-          s.Bounds = DiscoverBestBounds_MultipleVars(s.BoundVars, s.Range, true, ComprehensionExpr.BoundedPool.PoolVirtues.Enumerable);
-          s.BoundVars.Iter(bv => CheckTypeArgsContainNoOrdinal(bv.tok, bv.Type));
-
-        } else if (stmt is AssignSuchThatStmt) {
-          var s = (AssignSuchThatStmt)stmt;
-          if (s.AssumeToken == null) {
-            var varLhss = new List<IVariable>();
-            foreach (var lhs in s.Lhss) {
-              var ide = (IdentifierExpr)lhs.Resolved;  // successful resolution implies all LHS's are IdentifierExpr's
-              varLhss.Add(ide.Var);
-            }
-            s.Bounds = DiscoverBestBounds_MultipleVars(varLhss, s.Expr, true, ComprehensionExpr.BoundedPool.PoolVirtues.None);
-          }
-          foreach (var lhs in s.Lhss) {
-            var what = lhs is IdentifierExpr ? string.Format("variable '{0}'", ((IdentifierExpr)lhs).Name) : "LHS";
-            CheckTypeArgsContainNoOrdinal(lhs.tok, lhs.Type);
-          }
-        } else if (stmt is CalcStmt) {
-          var s = (CalcStmt)stmt;
-          // The resolution of the calc statement builds up .Steps and .Result, which are of the form E0 OP E1, where
-          // E0 and E1 are expressions from .Lines.  These additional expressions still need to have their .ResolvedOp
-          // fields filled in, so we visit them (but not their subexpressions) here.
-          foreach (var e in s.Steps) {
-            Visit(e);
-          }
-          Visit(s.Result);
-        }
-      }
-
-      protected override void VisitOneExpr(Expression expr) {
-        if (expr is LiteralExpr) {
-          var e = (LiteralExpr)expr;
-          if (e.Type.IsBitVectorType || e.Type.IsBigOrdinalType) {
-            var n = (BigInteger)e.Value;
-            var absN = n < 0 ? -n : n;
-            // For bitvectors, check that the magnitude fits the width
-            if (e.Type.IsBitVectorType && resolver.MaxBV(e.Type.AsBitVectorType.Width) < absN) {
-              resolver.reporter.Error(MessageSource.Resolver, e.tok, "literal ({0}) is too large for the bitvector type {1}", absN, e.Type);
-            }
-            // For bitvectors and ORDINALs, check for a unary minus that, earlier, was mistaken for a negative literal
-            // This can happen only in `match` patterns (see comment by LitPattern.OptimisticallyDesugaredLit).
-            if (n < 0 || e.tok.val == "-0") {
-              Contract.Assert(e.tok.val == "-0");  // this and the "if" above tests that "n < 0" happens only when the token is "-0"
-              resolver.reporter.Error(MessageSource.Resolver, e.tok, "unary minus (-{0}, type {1}) not allowed in case pattern", absN, e.Type);
-            }
-          }
-
-          if (expr is StaticReceiverExpr stexpr) {
-            if (stexpr.OriginalResolved != null) {
-              Visit(stexpr.OriginalResolved);
-            } else {
-              foreach (Type t in stexpr.Type.TypeArgs) {
-                if (t is InferredTypeProxy && ((InferredTypeProxy)t).T == null) {
-                  resolver.reporter.Error(MessageSource.Resolver, stexpr.tok, "type of type parameter could not be determined; please specify the type explicitly");
-                }
-              }
-            }
-          }
-
-        } else if (expr is ComprehensionExpr) {
-          var e = (ComprehensionExpr)expr;
-          foreach (var bv in e.BoundVars) {
-            if (!IsDetermined(bv.Type.Normalize())) {
-              resolver.reporter.Error(MessageSource.Resolver, bv.tok, "type of bound variable '{0}' could not be determined; please specify the type explicitly", bv.Name);
-            } else if (codeContext is ExtremePredicate) {
-              CheckContainsNoOrdinal(bv.tok, bv.Type, string.Format("type of bound variable '{0}' ('{1}') is not allowed to use type ORDINAL", bv.Name, bv.Type));
-            }
-          }
-          // apply bounds discovery to quantifiers, finite sets, and finite maps
-          string what = e.WhatKind;
-          Expression whereToLookForBounds = null;
-          var polarity = true;
-          if (e is QuantifierExpr quantifierExpr) {
-            whereToLookForBounds = quantifierExpr.LogicalBody();
-            polarity = quantifierExpr is ExistsExpr;
-          } else if (e is SetComprehension setComprehension) {
-            whereToLookForBounds = setComprehension.Range;
-          } else if (e is MapComprehension) {
-            whereToLookForBounds = e.Range;
-          } else {
-            Contract.Assume(e is LambdaExpr);  // otherwise, unexpected ComprehensionExpr
-          }
-          if (whereToLookForBounds != null) {
-            e.Bounds = DiscoverBestBounds_MultipleVars_AllowReordering(e.BoundVars, whereToLookForBounds, polarity, ComprehensionExpr.BoundedPool.PoolVirtues.None);
-            if (2 <= DafnyOptions.O.Allocated && (codeContext is Function || codeContext is ConstantField || CodeContextWrapper.Unwrap(codeContext) is RedirectingTypeDecl)) {
-              // functions are not allowed to depend on the set of allocated objects
-              foreach (var bv in ComprehensionExpr.BoundedPool.MissingBounds(e.BoundVars, e.Bounds, ComprehensionExpr.BoundedPool.PoolVirtues.IndependentOfAlloc)) {
-                var msgFormat = "a {0} involved in a {3} definition is not allowed to depend on the set of allocated references, but values of '{1}' may contain references";
-                if (bv.Type.IsTypeParameter || bv.Type.IsOpaqueType) {
-                  msgFormat += " (perhaps declare its type, '{2}', as '{2}(!new)')";
-                }
-                msgFormat += " (see documentation for 'older' parameters)";
-                var declKind = CodeContextWrapper.Unwrap(codeContext) is RedirectingTypeDecl redir ? redir.WhatKind : ((MemberDecl)codeContext).WhatKind;
-                resolver.reporter.Error(MessageSource.Resolver, e, msgFormat, what, bv.Name, bv.Type, declKind);
-              }
-            }
-            if ((e is SetComprehension && ((SetComprehension)e).Finite) || (e is MapComprehension && ((MapComprehension)e).Finite)) {
-              // the comprehension had better produce a finite set
-              if (e.Type.HasFinitePossibleValues) {
-                // This means the set is finite, regardless of if the Range is bounded.  So, we don't give any error here.
-                // However, if this expression is used in a non-ghost context (which is not yet known at this stage of
-                // resolution), the resolver will generate an error about that later.
-              } else {
-                // we cannot be sure that the set/map really is finite
-                foreach (var bv in ComprehensionExpr.BoundedPool.MissingBounds(e.BoundVars, e.Bounds, ComprehensionExpr.BoundedPool.PoolVirtues.Finite)) {
-                  resolver.reporter.Error(MessageSource.Resolver, e,
-                    "the result of a {0} must be finite, but Dafny's heuristics can't figure out how to produce a bounded set of values for '{1}'", what, bv.Name);
-                }
-              }
-            }
-          }
-
-          if (e is ExistsExpr && e.Range == null) {
-            var binBody = ((ExistsExpr)e).Term as BinaryExpr;
-            if (binBody != null && binBody.Op == BinaryExpr.Opcode.Imp) {  // check Op, not ResolvedOp, in order to distinguish ==> and <==
-              // apply the wisdom of Claude Marche: issue a warning here
-              resolver.reporter.Warning(MessageSource.Resolver, e.tok,
-                "the quantifier has the form 'exists x :: A ==> B', which most often is a typo for 'exists x :: A && B'; " +
-                "if you think otherwise, rewrite as 'exists x :: (A ==> B)' or 'exists x :: !A || B' to suppress this warning");
-            }
-          }
-
-        } else if (expr is MemberSelectExpr) {
-          var e = (MemberSelectExpr)expr;
-          if (e.Member is Function || e.Member is Method) {
-            var i = 0;
-            foreach (var p in Util.Concat(e.TypeApplication_AtEnclosingClass, e.TypeApplication_JustMember)) {
-              var tp = i < e.TypeApplication_AtEnclosingClass.Count ? e.Member.EnclosingClass.TypeArgs[i] : ((ICallable)e.Member).TypeArgs[i - e.TypeApplication_AtEnclosingClass.Count];
-              if (!IsDetermined(p.Normalize())) {
-                resolver.reporter.Error(MessageSource.Resolver, e.tok, "type parameter '{0}' (inferred to be '{1}') to the {2} '{3}' could not be determined", tp.Name, p, e.Member.WhatKind, e.Member.Name);
-              } else {
-                CheckContainsNoOrdinal(e.tok, p, string.Format("type parameter '{0}' (passed in as '{1}') to the {2} '{3}' is not allowed to use ORDINAL", tp.Name, p, e.Member.WhatKind, e.Member.Name));
-              }
-              i++;
-            }
-          }
-        } else if (expr is FunctionCallExpr) {
-          var e = (FunctionCallExpr)expr;
-          var i = 0;
-          foreach (var p in Util.Concat(e.TypeApplication_AtEnclosingClass, e.TypeApplication_JustFunction)) {
-            var tp = i < e.TypeApplication_AtEnclosingClass.Count ? e.Function.EnclosingClass.TypeArgs[i] : e.Function.TypeArgs[i - e.TypeApplication_AtEnclosingClass.Count];
-            if (!IsDetermined(p.Normalize())) {
-              resolver.reporter.Error(MessageSource.Resolver, e.tok, "type parameter '{0}' (inferred to be '{1}') in the function call to '{2}' could not be determined{3}", tp.Name, p, e.Name,
-                (e.Name.StartsWith("reveal_"))
-                ? ". If you are making an opaque function, make sure that the function can be called."
-                : ""
-              );
-            } else {
-              CheckContainsNoOrdinal(e.tok, p, string.Format("type parameter '{0}' (passed in as '{1}') to function call '{2}' is not allowed to use ORDINAL", tp.Name, p, e.Name));
-            }
-            i++;
-          }
-        } else if (expr is LetExpr) {
-          var e = (LetExpr)expr;
-          foreach (var p in e.LHSs) {
-            foreach (var x in p.Vars) {
-              if (!IsDetermined(x.Type.Normalize())) {
-                resolver.reporter.Error(MessageSource.Resolver, x.tok, "the type of the bound variable '{0}' could not be determined", x.Name);
-              } else {
-                CheckTypeArgsContainNoOrdinal(x.tok, x.Type);
-              }
-            }
-          }
-        } else if (expr is IdentifierExpr) {
-          // by specializing for IdentifierExpr, error messages will be clearer
-          CheckTypeIsDetermined(expr.tok, expr.Type, "variable");
-        } else if (expr is ConversionExpr) {
-          var e = (ConversionExpr)expr;
-          if (e.ToType.IsRefType) {
-            var fromType = e.E.Type;
-            Contract.Assert(fromType.IsRefType);
-            if (fromType.IsSubtypeOf(e.ToType, false, true) || e.ToType.IsSubtypeOf(fromType, false, true)) {
-              // looks good
-            } else {
-              resolver.reporter.Error(MessageSource.Resolver, e.tok,
-                "a type cast to a reference type ({0}) must be from a compatible type (got {1}); this cast could never succeed", e.ToType, fromType);
-            }
-          }
-        } else if (expr is TypeTestExpr) {
-          var e = (TypeTestExpr)expr;
-          var fromType = e.E.Type;
-          if (fromType.IsSubtypeOf(e.ToType, false, true)) {
-            // This test is allowed and it always returns true
-          } else if (!e.ToType.IsSubtypeOf(fromType, false, true)) {
-            resolver.reporter.Error(MessageSource.Resolver, e.tok,
-              "a type test to '{0}' must be from a compatible type (got '{1}')", e.ToType, fromType);
-          } else if (!e.ToType.IsRefType) {
-            resolver.reporter.Error(MessageSource.Resolver, e.tok,
-              "a non-trivial type test is allowed only for reference types (tried to test if '{1}' is a '{0}')", e.ToType, fromType);
-          }
-        } else if (CheckTypeIsDetermined(expr.tok, expr.Type, "expression")) {
-          if (expr is UnaryOpExpr uop) {
-            // The CheckTypeInference_Visitor has already visited uop.E, but uop.E's may be undetermined. If that happened,
-            // then an error has already been reported.
-            if (CheckTypeIsDetermined(uop.E.tok, uop.E.Type, "expression")) {
-              uop.ResolveOp(); // Force resolution eagerly at this point to catch potential bugs
-            }
-          } else if (expr is BinaryExpr) {
-            var e = (BinaryExpr)expr;
-            e.ResolvedOp = ResolveOp(e.Op, e.E0.Type, e.E1.Type);
-            // Check for useless comparisons with "null"
-            Expression other = null;  // if "null" if one of the operands, then "other" is the other
-            if (e.E0 is LiteralExpr && ((LiteralExpr)e.E0).Value == null) {
-              other = e.E1;
-            } else if (e.E1 is LiteralExpr && ((LiteralExpr)e.E1).Value == null) {
-              other = e.E0;
-            }
-            if (other != null) {
-              bool sense = true;
-              switch (e.ResolvedOp) {
-                case BinaryExpr.ResolvedOpcode.NeqCommon:
-                  sense = false;
-                  goto case BinaryExpr.ResolvedOpcode.EqCommon;
-                case BinaryExpr.ResolvedOpcode.EqCommon: {
-                    var nntUdf = other.Type.AsNonNullRefType;
-                    if (nntUdf != null) {
-                      string name = null;
-                      string hint = "";
-                      other = other.Resolved;
-                      if (other is IdentifierExpr) {
-                        name = string.Format("variable '{0}'", ((IdentifierExpr)other).Name);
-                      } else if (other is MemberSelectExpr) {
-                        var field = ((MemberSelectExpr)other).Member as Field;
-                        // The type of the field may be a formal type parameter, in which case the hint is omitted
-                        if (field.Type.IsNonNullRefType) {
-                          name = string.Format("field '{0}'", field.Name);
-                        }
-                      }
-                      if (name != null) {
-                        // The following relies on that a NonNullTypeDecl has a .Rhs that is a
-                        // UserDefinedType denoting the possibly null type declaration and that
-                        // these two types have the same number of type arguments.
-                        var nonNullTypeDecl = (NonNullTypeDecl)nntUdf.ResolvedClass;
-                        var possiblyNullUdf = (UserDefinedType)nonNullTypeDecl.Rhs;
-                        var possiblyNullTypeDecl = (ClassDecl)possiblyNullUdf.ResolvedClass;
-                        Contract.Assert(nonNullTypeDecl.TypeArgs.Count == possiblyNullTypeDecl.TypeArgs.Count);
-                        Contract.Assert(nonNullTypeDecl.TypeArgs.Count == nntUdf.TypeArgs.Count);
-                        var ty = new UserDefinedType(nntUdf.tok, possiblyNullUdf.Name, possiblyNullTypeDecl, nntUdf.TypeArgs);
-
-                        hint = string.Format(" (to make it possible for {0} to have the value 'null', declare its type to be '{1}')", name, ty);
-                      }
-                      resolver.reporter.Warning(MessageSource.Resolver, e.tok,
-                        string.Format("the type of the other operand is a non-null type, so this comparison with 'null' will always return '{0}'{1}",
-                        sense ? "false" : "true", hint));
-                    }
-                    break;
-                  }
-                case BinaryExpr.ResolvedOpcode.NotInSet:
-                case BinaryExpr.ResolvedOpcode.NotInSeq:
-                case BinaryExpr.ResolvedOpcode.NotInMultiSet:
-                  sense = false;
-                  goto case BinaryExpr.ResolvedOpcode.InSet;
-                case BinaryExpr.ResolvedOpcode.InSet:
-                case BinaryExpr.ResolvedOpcode.InSeq:
-                case BinaryExpr.ResolvedOpcode.InMultiSet: {
-                    var ty = other.Type.NormalizeExpand();
-                    var what = ty is SetType ? "set" : ty is SeqType ? "seq" : "multiset";
-                    if (((CollectionType)ty).Arg.IsNonNullRefType) {
-                      resolver.reporter.Warning(MessageSource.Resolver, e.tok,
-                        string.Format("the type of the other operand is a {0} of non-null elements, so the {1}inclusion test of 'null' will always return '{2}'",
-                        what, sense ? "" : "non-", sense ? "false" : "true"));
-                    }
-                    break;
-                  }
-                case BinaryExpr.ResolvedOpcode.NotInMap:
-                  goto case BinaryExpr.ResolvedOpcode.InMap;
-                case BinaryExpr.ResolvedOpcode.InMap: {
-                    var ty = other.Type.NormalizeExpand();
-                    if (((MapType)ty).Domain.IsNonNullRefType) {
-                      resolver.reporter.Warning(MessageSource.Resolver, e.tok,
-                        string.Format("the type of the other operand is a map to a non-null type, so the inclusion test of 'null' will always return '{0}'",
-                        sense ? "false" : "true"));
-                    }
-                    break;
-                  }
-                default:
-                  break;
-              }
-            }
-          } else if (expr is NegationExpression) {
-            var e = (NegationExpression)expr;
-            Expression resolved = null;
-            if (e.E is LiteralExpr lit) { // note, not e.E.Resolved, since we don't want to do this for double negations
-              // For real-based types, integer-based types, and bi (but not bitvectors), "-" followed by a literal is just a literal expression with a negative value
-              if (e.E.Type.IsNumericBased(Type.NumericPersuasion.Real)) {
-                var d = (BaseTypes.BigDec)lit.Value;
-                Contract.Assert(!d.IsNegative);
-                resolved = new LiteralExpr(e.tok, -d);
-              } else if (e.E.Type.IsNumericBased(Type.NumericPersuasion.Int)) {
-                var n = (BigInteger)lit.Value;
-                Contract.Assert(0 <= n);
-                resolved = new LiteralExpr(e.tok, -n);
-              }
-            }
-            if (resolved == null) {
-              // Treat all other expressions "-e" as "0 - e"
-              Expression zero;
-              if (e.E.Type.IsNumericBased(Type.NumericPersuasion.Real)) {
-                zero = new LiteralExpr(e.tok, BaseTypes.BigDec.ZERO);
-              } else {
-                Contract.Assert(e.E.Type.IsNumericBased(Type.NumericPersuasion.Int) || e.E.Type.IsBitVectorType);
-                zero = new LiteralExpr(e.tok, 0);
-              }
-              zero.Type = expr.Type;
-              resolved = new BinaryExpr(e.tok, BinaryExpr.Opcode.Sub, zero, e.E) { ResolvedOp = BinaryExpr.ResolvedOpcode.Sub };
-            }
-            resolved.Type = expr.Type;
-            e.ResolvedExpression = resolved;
-          }
-        }
-      }
-      public static bool IsDetermined(Type t) {
-        Contract.Requires(t != null);
-        Contract.Requires(!(t is TypeProxy) || ((TypeProxy)t).T == null);
-        // all other proxies indicate the type has not yet been determined, provided their type parameters have been
-        return !(t is TypeProxy) && t.TypeArgs.All(tt => IsDetermined(tt.Normalize()));
-      }
-      ISet<TypeProxy> UnderspecifiedTypeProxies = new HashSet<TypeProxy>();
-      public bool CheckTypeIsDetermined(IToken tok, Type t, string what) {
-        Contract.Requires(tok != null);
-        Contract.Requires(t != null);
-        Contract.Requires(what != null);
-        t = t.NormalizeExpand();
-
-        if (t is TypeProxy) {
-          var proxy = (TypeProxy)t;
-          if (!UnderspecifiedTypeProxies.Contains(proxy)) {
-            // report an error for this TypeProxy only once
-            resolver.reporter.Error(MessageSource.Resolver, tok, "the type of this {0} is underspecified", what);
-            UnderspecifiedTypeProxies.Add(proxy);
-          }
-          return false;
-        }
-        // Recurse on type arguments:
-        return t.TypeArgs.All(rg => CheckTypeIsDetermined(tok, rg, what));
-      }
-      public void CheckTypeArgsContainNoOrdinal(IToken tok, Type t) {
-        Contract.Requires(tok != null);
-        Contract.Requires(t != null);
-        t = t.NormalizeExpand();
-        t.TypeArgs.Iter(rg => CheckContainsNoOrdinal(tok, rg, "an ORDINAL type is not allowed to be used as a type argument"));
-      }
-      public void CheckContainsNoOrdinal(IToken tok, Type t, string errMsg) {
-        Contract.Requires(tok != null);
-        Contract.Requires(t != null);
-        Contract.Requires(errMsg != null);
-        t = t.NormalizeExpand();
-        if (t.IsBigOrdinalType) {
-          resolver.reporter.Error(MessageSource.Resolver, tok, errMsg);
-        }
-        t.TypeArgs.Iter(rg => CheckContainsNoOrdinal(tok, rg, errMsg));
-      }
-    }
-    #endregion CheckTypeInference
-
     /// <summary>
     /// Check that the 'older' modifier on parameters is used correctly and report any errors of the contrary.
     /// </summary>
@@ -7127,643 +6598,6 @@ namespace Microsoft.Dafny {
       }
     }
     #endregion
-
-    // ------------------------------------------------------------------------------------------------------
-    // ----- CheckTailRecursive -----------------------------------------------------------------------------
-    // ------------------------------------------------------------------------------------------------------
-    #region CheckTailRecursive
-    void DetermineTailRecursion(Method m) {
-      Contract.Requires(m != null);
-      Contract.Requires(m.Body != null);
-      bool tail = true;
-      bool hasTailRecursionPreference = Attributes.ContainsBool(m.Attributes, "tailrecursion", ref tail);
-      if (hasTailRecursionPreference && !tail) {
-        // the user specifically requested no tail recursion, so do nothing else
-      } else if (hasTailRecursionPreference && tail && m.IsGhost) {
-        reporter.Error(MessageSource.Resolver, m.tok, "tail recursion can be specified only for methods that will be compiled, not for ghost methods");
-      } else {
-        var module = m.EnclosingClass.EnclosingModuleDefinition;
-        var sccSize = module.CallGraph.GetSCCSize(m);
-        if (hasTailRecursionPreference && 2 <= sccSize) {
-          reporter.Error(MessageSource.Resolver, m.tok, "sorry, tail-call optimizations are not supported for mutually recursive methods");
-        } else if (hasTailRecursionPreference || sccSize == 1) {
-          Statement tailCall = null;
-          var status = CheckTailRecursive(m.Body.Body, m, ref tailCall, hasTailRecursionPreference);
-          if (status != TailRecursionStatus.NotTailRecursive && tailCall != null) {
-            // this means there was at least one recursive call
-            m.IsTailRecursive = true;
-            reporter.Info(MessageSource.Resolver, m.tok, "tail recursive");
-          }
-        }
-      }
-    }
-
-    enum TailRecursionStatus {
-      NotTailRecursive, // contains code that makes the enclosing method body not tail recursive (in way that is supported)
-      CanBeFollowedByAnything, // the code just analyzed does not do any recursive calls
-      TailCallSpent, // the method body is tail recursive, provided that all code that follows it in the method body is ghost
-    }
-
-    /// <summary>
-    /// Checks if "stmts" can be considered tail recursive, and (provided "reportError" is true) reports an error if not.
-    /// Note, the current implementation is rather conservative in its analysis; upon need, the
-    /// algorithm could be improved.
-    /// In the current implementation, "enclosingMethod" is not allowed to be a mutually recursive method.
-    ///
-    /// The incoming value of "tailCall" is not used, but it's nevertheless a 'ref' parameter to allow the
-    /// body to return the incoming value or to omit assignments to it.
-    /// If the return value is CanBeFollowedByAnything, "tailCall" is unchanged.
-    /// If the return value is TailCallSpent, "tailCall" shows one of the calls where the tail call was spent.  (Note,
-    /// there could be several if the statements have branches.)
-    /// If the return value is NoTailRecursive, "tailCall" could be anything.  In this case, an error
-    /// message has been reported (provided "reportsErrors" is true).
-    /// </summary>
-    TailRecursionStatus CheckTailRecursive(List<Statement> stmts, Method enclosingMethod, ref Statement tailCall, bool reportErrors) {
-      Contract.Requires(stmts != null);
-      var status = TailRecursionStatus.CanBeFollowedByAnything;
-      foreach (var s in stmts) {
-        if (!s.IsGhost) {
-          if (s is ReturnStmt && ((ReturnStmt)s).hiddenUpdate == null) {
-            return status;
-          }
-          if (status == TailRecursionStatus.TailCallSpent) {
-            // a tail call cannot be followed by non-ghost code
-            if (reportErrors) {
-              reporter.Error(MessageSource.Resolver, tailCall.Tok, "this recursive call is not recognized as being tail recursive, because it is followed by non-ghost code");
-            }
-            return TailRecursionStatus.NotTailRecursive;
-          }
-          status = CheckTailRecursive(s, enclosingMethod, ref tailCall, reportErrors);
-          if (status == TailRecursionStatus.NotTailRecursive) {
-            return status;
-          }
-        }
-      }
-      return status;
-    }
-
-    /// <summary>
-    /// See CheckTailRecursive(List Statement, ...), including its description of "tailCall".
-    /// In the current implementation, "enclosingMethod" is not allowed to be a mutually recursive method.
-    /// </summary>
-    TailRecursionStatus CheckTailRecursive(Statement stmt, Method enclosingMethod, ref Statement tailCall, bool reportErrors) {
-      Contract.Requires(stmt != null);
-      if (stmt.IsGhost) {
-        return TailRecursionStatus.CanBeFollowedByAnything;
-      }
-      if (stmt is PrintStmt) {
-      } else if (stmt is RevealStmt) {
-      } else if (stmt is BreakStmt) {
-      } else if (stmt is ReturnStmt) {
-        var s = (ReturnStmt)stmt;
-        if (s.hiddenUpdate != null) {
-          return CheckTailRecursive(s.hiddenUpdate, enclosingMethod, ref tailCall, reportErrors);
-        }
-      } else if (stmt is AssignStmt) {
-        var s = (AssignStmt)stmt;
-        if (s.Rhs is TypeRhs tRhs && tRhs.InitCall != null && tRhs.InitCall.Method == enclosingMethod) {
-          // It's a recursive call.  However, it is not a tail call, because after the "new" allocation
-          // and init call have taken place, the newly allocated object has yet to be assigned to
-          // the LHS of the assignment statement.
-          if (reportErrors) {
-            reporter.Error(MessageSource.Resolver, tRhs.InitCall.Tok,
-              "the recursive call to '{0}' is not tail recursive, because the assignment of the LHS happens after the call",
-              tRhs.InitCall.Method.Name);
-          }
-          return TailRecursionStatus.NotTailRecursive;
-        } else if (s.Rhs is ExprRhs eRhs && eRhs.Expr.Resolved is FunctionCallExpr fce && fce.Function.ByMethodDecl == enclosingMethod) {
-          var status = ConfirmTailCall(s.Tok, enclosingMethod, fce.TypeApplication_JustFunction, new List<Expression>() { s.Lhs }, reportErrors);
-          if (status == TailRecursionStatus.TailCallSpent) {
-            tailCall = s;
-            fce.Args.Iter(ee => DisallowRecursiveCallsInExpressions(ee, enclosingMethod));
-          } else {
-            DisallowRecursiveCallsInExpressions(s.Lhs, enclosingMethod, reportErrors);
-            DisallowRecursiveCallsInExpressions(eRhs.Expr, enclosingMethod, reportErrors);
-          }
-          return status;
-        }
-      } else if (stmt is ModifyStmt) {
-        var s = (ModifyStmt)stmt;
-        if (s.Body != null) {
-          return CheckTailRecursive(s.Body, enclosingMethod, ref tailCall, reportErrors);
-        }
-      } else if (stmt is CallStmt) {
-        var s = (CallStmt)stmt;
-        if (s.Method == enclosingMethod) {
-          DisallowRecursiveCallsInExpressions(s, enclosingMethod, reportErrors);
-          var status = ConfirmTailCall(s.Tok, s.Method, s.MethodSelect.TypeApplication_JustMember, s.Lhs, reportErrors);
-          if (status == TailRecursionStatus.TailCallSpent) {
-            tailCall = s;
-          }
-          return status;
-        }
-      } else if (stmt is BlockStmt) {
-        var s = (BlockStmt)stmt;
-        return CheckTailRecursive(s.Body, enclosingMethod, ref tailCall, reportErrors);
-      } else if (stmt is IfStmt) {
-        var s = (IfStmt)stmt;
-        DisallowRecursiveCallsInExpressions(s.Guard, enclosingMethod, reportErrors);
-        var stThen = CheckTailRecursive(s.Thn, enclosingMethod, ref tailCall, reportErrors);
-        if (stThen == TailRecursionStatus.NotTailRecursive) {
-          return stThen;
-        }
-        var stElse = s.Els == null ? TailRecursionStatus.CanBeFollowedByAnything : CheckTailRecursive(s.Els, enclosingMethod, ref tailCall, reportErrors);
-        if (stElse == TailRecursionStatus.NotTailRecursive) {
-          return stElse;
-        } else if (stThen == TailRecursionStatus.TailCallSpent || stElse == TailRecursionStatus.TailCallSpent) {
-          return TailRecursionStatus.TailCallSpent;
-        }
-      } else if (stmt is AlternativeStmt) {
-        var s = (AlternativeStmt)stmt;
-        var status = TailRecursionStatus.CanBeFollowedByAnything;
-        foreach (var alt in s.Alternatives) {
-          DisallowRecursiveCallsInExpressions(alt.Guard, enclosingMethod, reportErrors);
-          var st = CheckTailRecursive(alt.Body, enclosingMethod, ref tailCall, reportErrors);
-          if (st == TailRecursionStatus.NotTailRecursive) {
-            return st;
-          } else if (st == TailRecursionStatus.TailCallSpent) {
-            status = st;
-          }
-        }
-        return status;
-      } else if (stmt is OneBodyLoopStmt) {
-        var s = (OneBodyLoopStmt)stmt;
-        if (s is WhileStmt wh) {
-          DisallowRecursiveCallsInExpressions(wh.Guard, enclosingMethod, reportErrors);
-        } else if (s is ForLoopStmt forLoop) {
-          DisallowRecursiveCallsInExpressions(forLoop.Start, enclosingMethod, reportErrors);
-          DisallowRecursiveCallsInExpressions(forLoop.End, enclosingMethod, reportErrors);
-        }
-        var status = TailRecursionStatus.NotTailRecursive;
-        if (s.Body != null) {
-          status = CheckTailRecursive(s.Body, enclosingMethod, ref tailCall, reportErrors);
-        }
-        if (status != TailRecursionStatus.CanBeFollowedByAnything) {
-          if (status == TailRecursionStatus.NotTailRecursive) {
-            // an error has already been reported
-          } else if (reportErrors) {
-            reporter.Error(MessageSource.Resolver, tailCall.Tok, "a recursive call inside a loop is not recognized as being a tail call");
-          }
-          return TailRecursionStatus.NotTailRecursive;
-        }
-      } else if (stmt is AlternativeLoopStmt) {
-        var s = (AlternativeLoopStmt)stmt;
-        foreach (var alt in s.Alternatives) {
-          DisallowRecursiveCallsInExpressions(alt.Guard, enclosingMethod, reportErrors);
-          var status = CheckTailRecursive(alt.Body, enclosingMethod, ref tailCall, reportErrors);
-          if (status != TailRecursionStatus.CanBeFollowedByAnything) {
-            if (status == TailRecursionStatus.NotTailRecursive) {
-              // an error has already been reported
-            } else if (reportErrors) {
-              reporter.Error(MessageSource.Resolver, tailCall.Tok, "a recursive call inside a loop is not recognized as being a tail call");
-            }
-            return TailRecursionStatus.NotTailRecursive;
-          }
-        }
-      } else if (stmt is ForallStmt) {
-        var s = (ForallStmt)stmt;
-        DisallowRecursiveCallsInExpressions(s.Range, enclosingMethod, reportErrors);
-        var status = TailRecursionStatus.NotTailRecursive;
-        if (s.Body != null) {
-          status = CheckTailRecursive(s.Body, enclosingMethod, ref tailCall, reportErrors);
-        }
-        if (status != TailRecursionStatus.CanBeFollowedByAnything) {
-          if (status == TailRecursionStatus.NotTailRecursive) {
-            // an error has already been reported
-          } else if (reportErrors) {
-            reporter.Error(MessageSource.Resolver, tailCall.Tok, "a recursive call inside a forall statement is not a tail call");
-          }
-          return TailRecursionStatus.NotTailRecursive;
-        }
-      } else if (stmt is MatchStmt) {
-        var s = (MatchStmt)stmt;
-        DisallowRecursiveCallsInExpressions(s.Source, enclosingMethod, reportErrors);
-        var status = TailRecursionStatus.CanBeFollowedByAnything;
-        foreach (var kase in s.Cases) {
-          var st = CheckTailRecursive(kase.Body, enclosingMethod, ref tailCall, reportErrors);
-          if (st == TailRecursionStatus.NotTailRecursive) {
-            return st;
-          } else if (st == TailRecursionStatus.TailCallSpent) {
-            status = st;
-          }
-        }
-        return status;
-      } else if (stmt is ConcreteSyntaxStatement) {
-        var s = (ConcreteSyntaxStatement)stmt;
-        return CheckTailRecursive(s.ResolvedStatement, enclosingMethod, ref tailCall, reportErrors);
-      } else if (stmt is AssignSuchThatStmt) {
-      } else if (stmt is AssignOrReturnStmt) {
-        // TODO this should be the conservative choice, but probably we can consider this to be tail-recursive
-        // under some conditions? However, how does this interact with compiling to exceptions?
-        return TailRecursionStatus.NotTailRecursive;
-      } else if (stmt is UpdateStmt) {
-        var s = (UpdateStmt)stmt;
-        var ss = s.ResolvedStatements;
-        if (ss.Count == 1) {
-          return CheckTailRecursive(ss, enclosingMethod, ref tailCall, reportErrors);
-        } else {
-          foreach (var r in ss) {
-            DisallowRecursiveCallsInExpressions(r, enclosingMethod, reportErrors);
-          }
-        }
-      } else if (stmt is VarDeclStmt) {
-        var s = (VarDeclStmt)stmt;
-        if (s.Update != null) {
-          return CheckTailRecursive(s.Update, enclosingMethod, ref tailCall, reportErrors);
-        }
-      } else if (stmt is VarDeclPattern) {
-      } else if (stmt is ExpectStmt) {
-      } else {
-        Contract.Assert(false);  // unexpected statement type
-      }
-      DisallowRecursiveCallsInExpressions(stmt, enclosingMethod, reportErrors);
-      return TailRecursionStatus.CanBeFollowedByAnything;
-    }
-
-    /// <summary>
-    /// If "enclosingMethod" is a by-method, then look through "stmt" for any expression that
-    /// calls the function corresponding to the by-method. Report an error if such a call is
-    /// found.
-    /// </summary>
-    void DisallowRecursiveCallsInExpressions(Statement stmt, Method enclosingMethod, bool reportErrors) {
-      Contract.Requires(stmt != null);
-      Contract.Requires(enclosingMethod != null);
-
-      if (enclosingMethod.IsByMethod && reportErrors) {
-        stmt.SubExpressions.Iter(e => DisallowRecursiveCallsInExpressions(e, enclosingMethod));
-      }
-    }
-
-    void DisallowRecursiveCallsInExpressions(Expression/*?*/ expr, Method enclosingMethod, bool reportErrors) {
-      Contract.Requires(enclosingMethod != null);
-
-      if (expr != null && reportErrors) {
-        DisallowRecursiveCallsInExpressions(expr, enclosingMethod);
-      }
-    }
-
-    void DisallowRecursiveCallsInExpressions(Expression expr, Method enclosingMethod) {
-      Contract.Requires(expr != null);
-      Contract.Requires(enclosingMethod != null);
-
-      if (expr is FunctionCallExpr fce && fce.Function.ByMethodDecl == enclosingMethod) {
-        reporter.Error(MessageSource.Resolver, expr.tok, "a recursive call in this context is not recognized as a tail call");
-      }
-      expr.SubExpressions.Iter(ee => DisallowRecursiveCallsInExpressions(ee, enclosingMethod));
-    }
-
-    TailRecursionStatus ConfirmTailCall(IToken tok, Method method, List<Type> typeApplication_JustMember, List<Expression> lhss, bool reportErrors) {
-      Contract.Requires(tok != null);
-      Contract.Requires(method != null);
-      Contract.Requires(typeApplication_JustMember != null);
-      Contract.Requires(lhss != null);
-
-      // It's a recursive call.  It can be considered a tail call only if the LHS of the call are the
-      // formal out-parameters of the method
-      for (int i = 0; i < lhss.Count; i++) {
-        var formal = method.Outs[i];
-        if (!formal.IsGhost) {
-          if (lhss[i] is IdentifierExpr lhs && lhs.Var == formal) {
-            // all is good
-          } else {
-            if (reportErrors) {
-              var outParamName = formal.HasName ? $" '{formal.Name}'" : "";
-              reporter.Error(MessageSource.Resolver, tok,
-                "the recursive call to '{0}' is not tail recursive because the actual out-parameter{1} is not the formal out-parameter{2}",
-                method.Name, lhss.Count == 1 ? "" : " " + i, outParamName);
-            }
-            return TailRecursionStatus.NotTailRecursive;
-          }
-        }
-      }
-      // Moreover, it can be considered a tail recursive call only if the type parameters are the same
-      // as in the caller.
-      var classTypeParameterCount = method.EnclosingClass.TypeArgs.Count;
-      Contract.Assert(typeApplication_JustMember.Count == method.TypeArgs.Count);
-      for (int i = 0; i < method.TypeArgs.Count; i++) {
-        var formal = method.TypeArgs[i];
-        var actual = typeApplication_JustMember[i].AsTypeParameter;
-        if (formal != actual) {
-          if (reportErrors) {
-            reporter.Error(MessageSource.Resolver, tok,
-              "the recursive call to '{0}' is not tail recursive because the actual type parameter{1} is not the formal type parameter '{2}'",
-              method.Name, method.TypeArgs.Count == 1 ? "" : " " + i, formal.Name);
-          }
-          return TailRecursionStatus.NotTailRecursive;
-        }
-      }
-      return TailRecursionStatus.TailCallSpent;
-    }
-    #endregion CheckTailRecursive
-
-    // ------------------------------------------------------------------------------------------------------
-    // ----- CheckTailRecursiveExpr -------------------------------------------------------------------------
-    // ------------------------------------------------------------------------------------------------------
-    #region CheckTailRecursiveExpr
-    void DetermineTailRecursion(Function f) {
-      Contract.Requires(f != null);
-      Contract.Requires(f.Body != null);
-      bool tail = true;
-      bool hasTailRecursionPreference = Attributes.ContainsBool(f.Attributes, "tailrecursion", ref tail);
-      if (hasTailRecursionPreference && !tail) {
-        // the user specifically requested no tail recursion, so do nothing else
-      } else if (hasTailRecursionPreference && tail && f.IsGhost) {
-        reporter.Error(MessageSource.Resolver, f.tok, "tail recursion can be specified only for function that will be compiled, not for ghost functions");
-      } else {
-        var module = f.EnclosingClass.EnclosingModuleDefinition;
-        var sccSize = module.CallGraph.GetSCCSize(f);
-        if (hasTailRecursionPreference && 2 <= sccSize) {
-          reporter.Error(MessageSource.Resolver, f.tok, "sorry, tail-call optimizations are not supported for mutually recursive functions");
-        } else if (hasTailRecursionPreference || sccSize == 1) {
-          var status = CheckTailRecursiveExpr(f.Body, f, true, hasTailRecursionPreference);
-          if (status != Function.TailStatus.TriviallyTailRecursive && status != Function.TailStatus.NotTailRecursive) {
-            // this means there was at least one recursive call
-            f.TailRecursion = status;
-            if (status == Function.TailStatus.TailRecursive) {
-              reporter.Info(MessageSource.Resolver, f.tok, "tail recursive");
-            } else {
-              reporter.Info(MessageSource.Resolver, f.tok, "auto-accumulator tail recursive");
-            }
-          }
-        }
-      }
-    }
-
-    Function.TailStatus TRES_Or(Function.TailStatus a, Function.TailStatus b) {
-      if (a == Function.TailStatus.NotTailRecursive || b == Function.TailStatus.NotTailRecursive) {
-        return Function.TailStatus.NotTailRecursive;
-      } else if (a == Function.TailStatus.TriviallyTailRecursive) {
-        return b;
-      } else if (b == Function.TailStatus.TriviallyTailRecursive) {
-        return a;
-      } else if (a == Function.TailStatus.TailRecursive) {
-        return b;
-      } else if (b == Function.TailStatus.TailRecursive) {
-        return a;
-      } else if (a == b) {
-        return a;
-      } else {
-        return Function.TailStatus.NotTailRecursive;
-      }
-    }
-
-    /// <summary>
-    /// Checks if "expr" can be considered tail recursive, and (provided "reportError" is true) reports an error if not.
-    /// Note, the current implementation is rather conservative in its analysis; upon need, the
-    /// algorithm could be improved.
-    /// In the current implementation, "enclosingFunction" is not allowed to be a mutually recursive function.
-    ///
-    /// If "allowAccumulator" is "true", then tail recursion also allows expressions of the form "E * F"
-    /// and "F * E" where "F" is a tail-recursive expression without an accumulator, "E" has no occurrences
-    /// of the enclosing function, and "*" is an associative and eager operator with a known (left or right, respectively)
-    /// unit element. If "*" is such an operator, then "allowAccumulator" also allows expressions of
-    /// the form "F - E', where "-" is an operator that satisfies "(A - X) - Y == A - (X * Y)".
-    ///
-    /// If "allowAccumulator" is "false", then this method returns one of these three values:
-    ///     TriviallyTailRecursive, TailRecursive, NotTailRecursive
-    /// </summary>
-    Function.TailStatus CheckTailRecursiveExpr(Expression expr, Function enclosingFunction, bool allowAccumulator, bool reportErrors) {
-      Contract.Requires(expr != null);
-      Contract.Requires(enclosingFunction != null);
-
-      expr = expr.Resolved;
-      if (expr is FunctionCallExpr) {
-        var e = (FunctionCallExpr)expr;
-        var status = e.Function == enclosingFunction ? Function.TailStatus.TailRecursive : Function.TailStatus.TriviallyTailRecursive;
-        for (var i = 0; i < e.Function.Formals.Count; i++) {
-          if (!e.Function.Formals[i].IsGhost) {
-            var s = CheckHasNoRecursiveCall(e.Args[i], enclosingFunction, reportErrors);
-            status = TRES_Or(status, s);
-          }
-        }
-        return status;
-
-      } else if (expr is LetExpr) {
-        var e = (LetExpr)expr;
-        var status = Function.TailStatus.TriviallyTailRecursive;
-        for (var i = 0; i < e.LHSs.Count; i++) {
-          var pat = e.LHSs[i];
-          if (pat.Vars.ToList().Exists(bv => !bv.IsGhost)) {
-            if (e.Exact) {
-              var s = CheckHasNoRecursiveCall(e.RHSs[i], enclosingFunction, reportErrors);
-              status = TRES_Or(status, s);
-            } else {
-              // We have detected the existence of a non-ghost LHS, so check the RHS
-              Contract.Assert(e.RHSs.Count == 1);
-              status = CheckHasNoRecursiveCall(e.RHSs[0], enclosingFunction, reportErrors);
-              break;
-            }
-          }
-        }
-        var st = CheckTailRecursiveExpr(e.Body, enclosingFunction, allowAccumulator, reportErrors);
-        return TRES_Or(status, st);
-
-      } else if (expr is ITEExpr) {
-        var e = (ITEExpr)expr;
-        var s0 = CheckHasNoRecursiveCall(e.Test, enclosingFunction, reportErrors);
-        var s1 = CheckTailRecursiveExpr(e.Thn, enclosingFunction, allowAccumulator, reportErrors);
-        var s2 = CheckTailRecursiveExpr(e.Els, enclosingFunction, allowAccumulator, reportErrors);
-        var status = TRES_Or(s0, TRES_Or(s1, s2));
-        if (reportErrors && status == Function.TailStatus.NotTailRecursive) {
-          // We get here for one of the following reasons:
-          //   *  e.Test mentions the function (in which case an error has already been reported),
-          //   *  either e.Thn or e.Els was determined to be NotTailRecursive (in which case an
-          //      error has already been reported),
-          //   *  e.Thn and e.Els have different kinds of accumulator needs
-          if (s0 != Function.TailStatus.NotTailRecursive && s1 != Function.TailStatus.NotTailRecursive && s2 != Function.TailStatus.NotTailRecursive) {
-            reporter.Error(MessageSource.Resolver, expr, "if-then-else branches have different accumulator needs for tail recursion");
-          }
-        }
-        return status;
-
-      } else if (expr is MatchExpr) {
-        var e = (MatchExpr)expr;
-        var status = CheckHasNoRecursiveCall(e.Source, enclosingFunction, reportErrors);
-        var newError = reportErrors && status != Function.TailStatus.NotTailRecursive;
-        foreach (var kase in e.Cases) {
-          var s = CheckTailRecursiveExpr(kase.Body, enclosingFunction, allowAccumulator, reportErrors);
-          newError = newError && s != Function.TailStatus.NotTailRecursive;
-          status = TRES_Or(status, s);
-        }
-        if (status == Function.TailStatus.NotTailRecursive && newError) {
-          // see comments above for ITEExpr
-          // "newError" is "true" when: "reportErrors", and neither e.Source nor a kase.Body returned NotTailRecursive
-          reporter.Error(MessageSource.Resolver, expr, "cases have different accumulator needs for tail recursion");
-        }
-        return status;
-
-      } else if (allowAccumulator && expr is BinaryExpr bin) {
-        var accumulationOp = Function.TailStatus.TriviallyTailRecursive; // use TriviallyTailRecursive to mean bin.ResolvedOp does not support accumulation
-        bool accumulatesOnlyOnRight = false;
-        switch (bin.ResolvedOp) {
-          case BinaryExpr.ResolvedOpcode.Add:
-            if (enclosingFunction.ResultType.AsBitVectorType == null && !enclosingFunction.ResultType.IsCharType) {
-              accumulationOp = Function.TailStatus.Accumulate_Add;
-            }
-            break;
-          case BinaryExpr.ResolvedOpcode.Sub:
-            if (enclosingFunction.ResultType.AsBitVectorType == null && !enclosingFunction.ResultType.IsCharType) {
-              accumulationOp = Function.TailStatus.AccumulateRight_Sub;
-              accumulatesOnlyOnRight = true;
-            }
-            break;
-          case BinaryExpr.ResolvedOpcode.Mul:
-            if (enclosingFunction.ResultType.AsBitVectorType == null) {
-              accumulationOp = Function.TailStatus.Accumulate_Mul;
-            }
-            break;
-          case BinaryExpr.ResolvedOpcode.Union:
-            accumulationOp = Function.TailStatus.Accumulate_SetUnion;
-            break;
-          case BinaryExpr.ResolvedOpcode.SetDifference:
-            accumulationOp = Function.TailStatus.AccumulateRight_SetDifference;
-            accumulatesOnlyOnRight = true;
-            break;
-          case BinaryExpr.ResolvedOpcode.MultiSetUnion:
-            accumulationOp = Function.TailStatus.Accumulate_MultiSetUnion;
-            break;
-          case BinaryExpr.ResolvedOpcode.MultiSetDifference:
-            accumulationOp = Function.TailStatus.AccumulateRight_MultiSetDifference;
-            accumulatesOnlyOnRight = true;
-            break;
-          case BinaryExpr.ResolvedOpcode.Concat:
-            accumulationOp = Function.TailStatus.AccumulateLeft_Concat;  // could also be AccumulateRight_Concat--make more precise below
-            break;
-          default:
-            break;
-        }
-        if (accumulationOp != Function.TailStatus.TriviallyTailRecursive) {
-          var s0 = CheckTailRecursiveExpr(bin.E0, enclosingFunction, false, reportErrors);
-          Function.TailStatus s1;
-          switch (s0) {
-            case Function.TailStatus.NotTailRecursive:
-              // Any errors have already been reported, but still descend down bin.E1 (possibly reporting
-              // more errors) before returning with NotTailRecursive
-              s1 = CheckTailRecursiveExpr(bin.E1, enclosingFunction, false, reportErrors);
-              return s0;
-            case Function.TailStatus.TriviallyTailRecursive:
-              // We are in a state that would allow AcculumateLeftTailRecursive. See what bin.E1 is like:
-              if (accumulatesOnlyOnRight) {
-                s1 = CheckHasNoRecursiveCall(bin.E1, enclosingFunction, reportErrors);
-              } else {
-                s1 = CheckTailRecursiveExpr(bin.E1, enclosingFunction, false, reportErrors);
-              }
-              if (s1 == Function.TailStatus.TailRecursive) {
-                bin.AccumulatesForTailRecursion = BinaryExpr.AccumulationOperand.Left;
-              } else {
-                Contract.Assert(s1 == Function.TailStatus.TriviallyTailRecursive || s1 == Function.TailStatus.NotTailRecursive);
-                return s1;
-              }
-              return accumulationOp;
-            case Function.TailStatus.TailRecursive:
-              // We are in a state that would allow right-accumulative tail recursion. Check that the enclosing
-              // function is not mentioned in bin.E1.
-              s1 = CheckHasNoRecursiveCall(bin.E1, enclosingFunction, reportErrors);
-              if (s1 == Function.TailStatus.TriviallyTailRecursive) {
-                bin.AccumulatesForTailRecursion = BinaryExpr.AccumulationOperand.Right;
-                if (accumulationOp == Function.TailStatus.AccumulateLeft_Concat) {
-                  // switch to AccumulateRight_Concat, since we had approximated it as AccumulateLeft_Concat above
-                  return Function.TailStatus.AccumulateRight_Concat;
-                } else {
-                  return accumulationOp;
-                }
-              } else {
-                Contract.Assert(s1 == Function.TailStatus.NotTailRecursive);
-                return s1;
-              }
-            default:
-              Contract.Assert(false); // unexpected case
-              throw new cce.UnreachableException();
-          }
-        }
-        // not an operator that allows accumulation, so drop down below
-      } else if (expr is StmtExpr) {
-        var e = (StmtExpr)expr;
-        // ignore the statement part, since it is ghost
-        return CheckTailRecursiveExpr(e.E, enclosingFunction, allowAccumulator, reportErrors);
-      }
-
-      return CheckHasNoRecursiveCall(expr, enclosingFunction, reportErrors);
-    }
-
-    /// <summary>
-    /// If "expr" contains a recursive call to "enclosingFunction" in some non-ghost subexpressions,
-    /// then returns TailStatus.NotTailRecursive (and if "reportErrors" is "true", then
-    /// reports an error about the recursive call), else returns TailStatus.TriviallyTailRecursive.
-    /// </summary>
-    Function.TailStatus CheckHasNoRecursiveCall(Expression expr, Function enclosingFunction, bool reportErrors) {
-      Contract.Requires(expr != null);
-      Contract.Requires(enclosingFunction != null);
-
-      var status = Function.TailStatus.TriviallyTailRecursive;
-
-      if (expr is FunctionCallExpr) {
-        var e = (FunctionCallExpr)expr;
-        if (e.Function == enclosingFunction) {
-          if (reportErrors) {
-            reporter.Error(MessageSource.Resolver, expr, "to be tail recursive, every use of this function must be part of a tail call or a simple accumulating tail call");
-          }
-          status = Function.TailStatus.NotTailRecursive;
-        }
-        // skip ghost subexpressions
-        for (var i = 0; i < e.Function.Formals.Count; i++) {
-          if (!e.Function.Formals[i].IsGhost) {
-            var s = CheckHasNoRecursiveCall(e.Args[i], enclosingFunction, reportErrors);
-            status = TRES_Or(status, s);
-          }
-        }
-        return status;
-
-      } else if (expr is MemberSelectExpr) {
-        var e = (MemberSelectExpr)expr;
-        if (e.Member == enclosingFunction) {
-          if (reportErrors) {
-            reporter.Error(MessageSource.Resolver, expr, "to be tail recursive, every use of this function must be part of a tail call or a simple accumulating tail call");
-          }
-          return Function.TailStatus.NotTailRecursive;
-        }
-
-      } else if (expr is LetExpr) {
-        var e = (LetExpr)expr;
-        // skip ghost subexpressions
-        for (var i = 0; i < e.LHSs.Count; i++) {
-          var pat = e.LHSs[i];
-          if (pat.Vars.ToList().Exists(bv => !bv.IsGhost)) {
-            if (e.Exact) {
-              var s = CheckHasNoRecursiveCall(e.RHSs[i], enclosingFunction, reportErrors);
-              status = TRES_Or(status, s);
-            } else {
-              // We have detected the existence of a non-ghost LHS, so check the RHS
-              Contract.Assert(e.RHSs.Count == 1);
-              status = CheckHasNoRecursiveCall(e.RHSs[0], enclosingFunction, reportErrors);
-              break;
-            }
-          }
-        }
-        var st = CheckHasNoRecursiveCall(e.Body, enclosingFunction, reportErrors);
-        return TRES_Or(status, st);
-
-      } else if (expr is DatatypeValue) {
-        var e = (DatatypeValue)expr;
-        // skip ghost subexpressions
-        for (var i = 0; i < e.Ctor.Formals.Count; i++) {
-          if (!e.Ctor.Formals[i].IsGhost) {
-            var s = CheckHasNoRecursiveCall(e.Arguments[i], enclosingFunction, reportErrors);
-            status = TRES_Or(status, s);
-          }
-        }
-        return status;
-
-      }
-
-      foreach (var ee in expr.SubExpressions) {
-        var s = CheckHasNoRecursiveCall(ee, enclosingFunction, reportErrors);
-        status = TRES_Or(status, s);
-      }
-      return status;
-    }
-
-    #endregion CheckTailRecursiveExpr
 
     // ------------------------------------------------------------------------------------------------------
     // ----- FuelAdjustmentChecks ---------------------------------------------------------------------------
@@ -8459,547 +7293,7 @@ namespace Microsoft.Dafny {
       var visitor = new GhostInterest_Visitor(codeContext, this, false);
       visitor.Visit(stmt, mustBeErasable, proofContext);
     }
-    class GhostInterest_Visitor {
-      readonly ICodeContext codeContext;
-      readonly Resolver resolver;
-      private readonly bool inConstructorInitializationPhase;
-      public GhostInterest_Visitor(ICodeContext codeContext, Resolver resolver, bool inConstructorInitializationPhase) {
-        Contract.Requires(codeContext != null);
-        Contract.Requires(resolver != null);
-        this.codeContext = codeContext;
-        this.resolver = resolver;
-        this.inConstructorInitializationPhase = inConstructorInitializationPhase;
-      }
-      protected void Error(Statement stmt, string msg, params object[] msgArgs) {
-        Contract.Requires(stmt != null);
-        Contract.Requires(msg != null);
-        Contract.Requires(msgArgs != null);
-        resolver.reporter.Error(MessageSource.Resolver, stmt, msg, msgArgs);
-      }
-      protected void Error(Expression expr, string msg, params object[] msgArgs) {
-        Contract.Requires(expr != null);
-        Contract.Requires(msg != null);
-        Contract.Requires(msgArgs != null);
-        resolver.reporter.Error(MessageSource.Resolver, expr, msg, msgArgs);
-      }
-      protected void Error(IToken tok, string msg, params object[] msgArgs) {
-        Contract.Requires(tok != null);
-        Contract.Requires(msg != null);
-        Contract.Requires(msgArgs != null);
-        resolver.reporter.Error(MessageSource.Resolver, tok, msg, msgArgs);
-      }
-      /// <summary>
-      /// There are three kinds of contexts for statements.
-      ///   - compiled contexts, where the statement must be compilable
-      ///     -- !mustBeErasable && proofContext == null
-      ///   - ghost contexts that allow the allocation of new object
-      ///     -- mustBeErasable && proofContext == null
-      ///   - lemma/proof contexts, which are ghost and are not allowed to allocate new objects
-      ///     -- mustBeErasable && proofContext != null
-      /// 
-      /// This method does three things, in order:
-      /// 0. Sets .IsGhost to "true" if the statement is ghost.  This often depends on some guard of the statement
-      ///    (like the guard of an "if" statement) or the LHS of the statement (if it is an assignment).
-      ///    Note, if "mustBeErasable", then the statement is already in a ghost context.
-      /// 1. Determines if the statement and all its subparts are legal under its computed .IsGhost setting.
-      /// 2. ``Upgrades'' .IsGhost to "true" if, after investigation of the substatements of the statement, it
-      ///    turns out that the statement can be erased during compilation.
-      /// Notes:
-      /// * Both step (0) and step (2) sets the .IsGhost field.  What step (0) does affects only the
-      ///   rules of resolution, whereas step (2) makes a note for the later compilation phase.
-      /// * It is important to do step (0) before step (1)--that is, it is important to set the statement's ghost
-      ///   status before descending into its substatements--because break statements look at the ghost status of
-      ///   its enclosing statements.
-      /// * The method called by a StmtExpr must be ghost; however, this is checked elsewhere.  For
-      ///   this reason, it is not necessary to visit all subexpressions, unless the subexpression
-      ///   matter for the ghost checking/recording of "stmt".
-      ///
-      /// If "proofContext" is non-null, then this method also checks that "stmt" does not allocate
-      /// memory or modify the heap, either directly or indirectly using a statement like "modify", a loop with
-      /// an explicit "modifies" clause, or a call to a method that may allocate memory or modify the heap.
-      /// The "proofContext" string is something that can be printed as part of an error message.
-      /// </summary>
-      public void Visit(Statement stmt, bool mustBeErasable, [CanBeNull] string proofContext) {
-        Contract.Requires(stmt != null);
-        Contract.Assume(!codeContext.IsGhost || mustBeErasable); // (this is really a precondition) CodeContext.IsGhost ==> mustBeErasable
-        Contract.Assume(mustBeErasable || proofContext == null); // (this is really a precondition) !mustBeErasable ==> proofContext == null 
 
-        if (stmt is AssertStmt || stmt is AssumeStmt) {
-          stmt.IsGhost = true;
-          var assertStmt = stmt as AssertStmt;
-          if (assertStmt != null && assertStmt.Proof != null) {
-            Visit(assertStmt.Proof, true, "an assert-by body");
-          }
-
-        } else if (stmt is ExpectStmt) {
-          stmt.IsGhost = false;
-          var s = (ExpectStmt)stmt;
-          if (mustBeErasable) {
-            Error(stmt, "expect statement is not allowed in this context (because this is a ghost method or because the statement is guarded by a specification-only expression)");
-          } else {
-            ExpressionTester.CheckIsCompilable(resolver, s.Expr, codeContext);
-            // If not provided, the message is populated with a default value in resolution
-            Contract.Assert(s.Message != null);
-            ExpressionTester.CheckIsCompilable(resolver, s.Message, codeContext);
-          }
-
-        } else if (stmt is PrintStmt) {
-          var s = (PrintStmt)stmt;
-          if (mustBeErasable) {
-            Error(stmt, "print statement is not allowed in this context (because this is a ghost method or because the statement is guarded by a specification-only expression)");
-          } else {
-            s.Args.Iter(ee => ExpressionTester.CheckIsCompilable(resolver, ee, codeContext));
-          }
-
-        } else if (stmt is RevealStmt) {
-          var s = (RevealStmt)stmt;
-          s.ResolvedStatements.Iter(ss => Visit(ss, true, "a reveal statement"));
-          s.IsGhost = s.ResolvedStatements.All(ss => ss.IsGhost);
-
-        } else if (stmt is BreakStmt) {
-          var s = (BreakStmt)stmt;
-          s.IsGhost = mustBeErasable;
-          if (s.IsGhost && !s.TargetStmt.IsGhost) {
-            var targetKind = s.TargetStmt is LoopStmt ? "loop" : "structure";
-            Error(stmt, $"ghost-context {s.Kind} statement is not allowed to {s.Kind} out of non-ghost {targetKind}");
-          }
-
-        } else if (stmt is ProduceStmt) {
-          var s = (ProduceStmt)stmt;
-          var kind = stmt is YieldStmt ? "yield" : "return";
-          if (mustBeErasable && !codeContext.IsGhost) {
-            Error(stmt, "{0} statement is not allowed in this context (because it is guarded by a specification-only expression)", kind);
-          }
-          if (s.hiddenUpdate != null) {
-            Visit(s.hiddenUpdate, mustBeErasable, proofContext);
-          }
-
-        } else if (stmt is AssignSuchThatStmt) {
-          var s = (AssignSuchThatStmt)stmt;
-          s.IsGhost = mustBeErasable || s.AssumeToken != null || s.Lhss.Any(AssignStmt.LhsIsToGhost);
-          if (mustBeErasable && !codeContext.IsGhost) {
-            foreach (var lhs in s.Lhss) {
-              var gk = AssignStmt.LhsIsToGhost_Which(lhs);
-              if (gk != AssignStmt.NonGhostKind.IsGhost) {
-                Error(lhs, "cannot assign to {0} in a ghost context", AssignStmt.NonGhostKind_To_String(gk));
-              }
-            }
-          } else if (!mustBeErasable && s.AssumeToken == null && ExpressionTester.UsesSpecFeatures(s.Expr)) {
-            foreach (var lhs in s.Lhss) {
-              var gk = AssignStmt.LhsIsToGhost_Which(lhs);
-              if (gk != AssignStmt.NonGhostKind.IsGhost) {
-                Error(lhs, "{0} cannot be assigned a value that depends on a ghost", AssignStmt.NonGhostKind_To_String(gk));
-              }
-            }
-          }
-
-        } else if (stmt is UpdateStmt) {
-          var s = (UpdateStmt)stmt;
-          s.ResolvedStatements.Iter(ss => Visit(ss, mustBeErasable, proofContext));
-          s.IsGhost = s.ResolvedStatements.All(ss => ss.IsGhost);
-
-        } else if (stmt is AssignOrReturnStmt) {
-          var s = (AssignOrReturnStmt)stmt;
-          s.ResolvedStatements.Iter(ss => Visit(ss, mustBeErasable, proofContext));
-          s.IsGhost = s.ResolvedStatements.All(ss => ss.IsGhost);
-
-        } else if (stmt is VarDeclStmt) {
-          var s = (VarDeclStmt)stmt;
-          if (mustBeErasable) {
-            foreach (var local in s.Locals) {
-              // a local variable in a specification-only context might as well be ghost
-              local.MakeGhost();
-            }
-          }
-          if (s.Update != null) {
-            Visit(s.Update, mustBeErasable, proofContext);
-          }
-          s.IsGhost = (s.Update == null || s.Update.IsGhost) && s.Locals.All(v => v.IsGhost);
-
-        } else if (stmt is VarDeclPattern) {
-          var s = (VarDeclPattern)stmt;
-
-          if (mustBeErasable) {
-            foreach (var local in s.LocalVars) {
-              local.MakeGhost();
-            }
-          }
-          if (s.HasGhostModifier || mustBeErasable) {
-            s.IsGhost = s.LocalVars.All(v => v.IsGhost);
-          } else {
-            var spec = ExpressionTester.UsesSpecFeatures(s.RHS);
-            if (spec) {
-              foreach (var local in s.LocalVars) {
-                local.MakeGhost();
-              }
-            } else {
-              ExpressionTester.CheckIsCompilable(resolver, s.RHS, codeContext);
-            }
-            s.IsGhost = spec;
-          }
-
-        } else if (stmt is AssignStmt) {
-          var s = (AssignStmt)stmt;
-          CheckAssignStmt(s, mustBeErasable, proofContext);
-
-        } else if (stmt is CallStmt) {
-          var s = (CallStmt)stmt;
-          var callee = s.Method;
-          Contract.Assert(callee != null);  // follows from the invariant of CallStmt
-          s.IsGhost = callee.IsGhost;
-          if (proofContext != null && !callee.IsLemmaLike) {
-            Error(s, $"in {proofContext}, calls are allowed only to lemmas");
-          } else if (mustBeErasable) {
-            if (!s.IsGhost) {
-              Error(s, "only ghost methods can be called from this context");
-            }
-          } else {
-            int j;
-            if (!callee.IsGhost) {
-              // check in-parameters
-              ExpressionTester.CheckIsCompilable(resolver, s.Receiver, codeContext);
-              j = 0;
-              foreach (var e in s.Args) {
-                Contract.Assume(j < callee.Ins.Count);  // this should have already been checked by the resolver
-                if (!callee.Ins[j].IsGhost) {
-                  ExpressionTester.CheckIsCompilable(resolver, e, codeContext);
-                }
-                j++;
-              }
-            }
-            j = 0;
-            foreach (var e in s.Lhs) {
-              var resolvedLhs = e.Resolved;
-              if (callee.IsGhost || callee.Outs[j].IsGhost) {
-                // LHS must denote a ghost
-                if (resolvedLhs is IdentifierExpr) {
-                  var ll = (IdentifierExpr)resolvedLhs;
-                  if (!ll.Var.IsGhost) {
-                    if (ll is AutoGhostIdentifierExpr && ll.Var is LocalVariable) {
-                      // the variable was actually declared in this statement, so auto-declare it as ghost
-                      ((LocalVariable)ll.Var).MakeGhost();
-                    } else {
-                      Error(s, "actual out-parameter{0} is required to be a ghost variable", s.Lhs.Count == 1 ? "" : " " + j);
-                    }
-                  }
-                } else if (resolvedLhs is MemberSelectExpr) {
-                  var ll = (MemberSelectExpr)resolvedLhs;
-                  if (!ll.Member.IsGhost) {
-                    Error(s, "actual out-parameter{0} is required to be a ghost field", s.Lhs.Count == 1 ? "" : " " + j);
-                  }
-                } else {
-                  // this is an array update, and arrays are always non-ghost
-                  Error(s, "actual out-parameter{0} is required to be a ghost variable", s.Lhs.Count == 1 ? "" : " " + j);
-                }
-              }
-              j++;
-            }
-          }
-
-        } else if (stmt is BlockStmt) {
-          var s = (BlockStmt)stmt;
-          s.IsGhost = mustBeErasable;  // set .IsGhost before descending into substatements (since substatements may do a 'break' out of this block)
-          if (s is DividedBlockStmt ds) {
-            var giv = new GhostInterest_Visitor(this.codeContext, this.resolver, true);
-            ds.BodyInit.Iter(ss => giv.Visit(ss, mustBeErasable, proofContext));
-            ds.BodyProper.Iter(ss => Visit(ss, mustBeErasable, proofContext));
-          } else {
-            s.Body.Iter(ss => Visit(ss, mustBeErasable, proofContext));
-          }
-          s.IsGhost = s.IsGhost || s.Body.All(ss => ss.IsGhost);  // mark the block statement as ghost if all its substatements are ghost
-
-        } else if (stmt is IfStmt) {
-          var s = (IfStmt)stmt;
-          s.IsGhost = mustBeErasable || (s.Guard != null && ExpressionTester.UsesSpecFeatures(s.Guard));
-          if (!mustBeErasable && s.IsGhost) {
-            resolver.reporter.Info(MessageSource.Resolver, s.Tok, "ghost if");
-          }
-          Visit(s.Thn, s.IsGhost, proofContext);
-          if (s.Els != null) {
-            Visit(s.Els, s.IsGhost, proofContext);
-          }
-          // if both branches were all ghost, then we can mark the enclosing statement as ghost as well
-          s.IsGhost = s.IsGhost || (s.Thn.IsGhost && (s.Els == null || s.Els.IsGhost));
-          if (!s.IsGhost && s.Guard != null) {
-            // If there were features in the guard that are treated differently in ghost and non-ghost
-            // contexts, make sure they get treated for non-ghost use.
-            ExpressionTester.CheckIsCompilable(resolver, s.Guard, codeContext);
-          }
-
-        } else if (stmt is AlternativeStmt) {
-          var s = (AlternativeStmt)stmt;
-          s.IsGhost = mustBeErasable || s.Alternatives.Exists(alt => ExpressionTester.UsesSpecFeatures(alt.Guard));
-          if (!mustBeErasable && s.IsGhost) {
-            resolver.reporter.Info(MessageSource.Resolver, s.Tok, "ghost if");
-          }
-          s.Alternatives.Iter(alt => alt.Body.Iter(ss => Visit(ss, s.IsGhost, proofContext)));
-          s.IsGhost = s.IsGhost || s.Alternatives.All(alt => alt.Body.All(ss => ss.IsGhost));
-          if (!s.IsGhost) {
-            // If there were features in the guards that are treated differently in ghost and non-ghost
-            // contexts, make sure they get treated for non-ghost use.
-            foreach (var alt in s.Alternatives) {
-              ExpressionTester.CheckIsCompilable(resolver, alt.Guard, codeContext);
-            }
-          }
-
-        } else if (stmt is WhileStmt) {
-          var s = (WhileStmt)stmt;
-          if (proofContext != null && s.Mod.Expressions != null && s.Mod.Expressions.Count != 0) {
-            Error(s.Mod.Expressions[0].tok, $"a loop in {proofContext} is not allowed to use 'modifies' clauses");
-          }
-
-          s.IsGhost = mustBeErasable || (s.Guard != null && ExpressionTester.UsesSpecFeatures(s.Guard));
-          if (!mustBeErasable && s.IsGhost) {
-            resolver.reporter.Info(MessageSource.Resolver, s.Tok, "ghost while");
-          }
-          if (s.IsGhost && s.Decreases.Expressions.Exists(e => e is WildcardExpr)) {
-            Error(s, "'decreases *' is not allowed on ghost loops");
-          }
-          if (s.IsGhost && s.Mod.Expressions != null) {
-            s.Mod.Expressions.Iter(resolver.DisallowNonGhostFieldSpecifiers);
-          }
-          if (s.Body != null) {
-            Visit(s.Body, s.IsGhost, proofContext);
-            if (s.Body.IsGhost && !s.Decreases.Expressions.Exists(e => e is WildcardExpr)) {
-              s.IsGhost = true;
-            }
-          }
-          if (!s.IsGhost && s.Guard != null) {
-            // If there were features in the guard that are treated differently in ghost and non-ghost
-            // contexts, make sure they get treated for non-ghost use.
-            ExpressionTester.CheckIsCompilable(resolver, s.Guard, codeContext);
-          }
-
-        } else if (stmt is AlternativeLoopStmt) {
-          var s = (AlternativeLoopStmt)stmt;
-          if (proofContext != null && s.Mod.Expressions != null && s.Mod.Expressions.Count != 0) {
-            Error(s.Mod.Expressions[0].tok, $"a loop in {proofContext} is not allowed to use 'modifies' clauses");
-          }
-
-          s.IsGhost = mustBeErasable || s.Alternatives.Exists(alt => ExpressionTester.UsesSpecFeatures(alt.Guard));
-          if (!mustBeErasable && s.IsGhost) {
-            resolver.reporter.Info(MessageSource.Resolver, s.Tok, "ghost while");
-          }
-          if (s.IsGhost && s.Decreases.Expressions.Exists(e => e is WildcardExpr)) {
-            Error(s, "'decreases *' is not allowed on ghost loops");
-          }
-          if (s.IsGhost && s.Mod.Expressions != null) {
-            s.Mod.Expressions.Iter(resolver.DisallowNonGhostFieldSpecifiers);
-          }
-          s.Alternatives.Iter(alt => alt.Body.Iter(ss => Visit(ss, s.IsGhost, proofContext)));
-          s.IsGhost = s.IsGhost || (!s.Decreases.Expressions.Exists(e => e is WildcardExpr) && s.Alternatives.All(alt => alt.Body.All(ss => ss.IsGhost)));
-          if (!s.IsGhost) {
-            // If there were features in the guards that are treated differently in ghost and non-ghost
-            // contexts, make sure they get treated for non-ghost use.
-            foreach (var alt in s.Alternatives) {
-              ExpressionTester.CheckIsCompilable(resolver, alt.Guard, codeContext);
-            }
-          }
-
-        } else if (stmt is ForLoopStmt) {
-          var s = (ForLoopStmt)stmt;
-          if (proofContext != null && s.Mod.Expressions != null && s.Mod.Expressions.Count != 0) {
-            Error(s.Mod.Expressions[0].tok, $"a loop in {proofContext} is not allowed to use 'modifies' clauses");
-          }
-
-          s.IsGhost = mustBeErasable || ExpressionTester.UsesSpecFeatures(s.Start) || (s.End != null && ExpressionTester.UsesSpecFeatures(s.End));
-          if (!mustBeErasable && s.IsGhost) {
-            resolver.reporter.Info(MessageSource.Resolver, s.Tok, "ghost for-loop");
-          }
-          if (s.IsGhost) {
-            if (s.Decreases.Expressions.Exists(e => e is WildcardExpr)) {
-              Error(s, "'decreases *' is not allowed on ghost loops");
-            } else if (s.End == null && s.Decreases.Expressions.Count == 0) {
-              Error(s, "a ghost loop must be terminating; make the end-expression specific or add a 'decreases' clause");
-            }
-          }
-          if (s.IsGhost && s.Mod.Expressions != null) {
-            s.Mod.Expressions.Iter(resolver.DisallowNonGhostFieldSpecifiers);
-          }
-          if (s.Body != null) {
-            Visit(s.Body, s.IsGhost, proofContext);
-            if (s.Body.IsGhost) {
-              s.IsGhost = true;
-            }
-          }
-          if (!s.IsGhost) {
-            // If there were features in the bounds that are treated differently in ghost and non-ghost
-            // contexts, make sure they get treated for non-ghost use.
-            ExpressionTester.CheckIsCompilable(resolver, s.Start, codeContext);
-            if (s.End != null) {
-              ExpressionTester.CheckIsCompilable(resolver, s.End, codeContext);
-            }
-          }
-
-        } else if (stmt is ForallStmt) {
-          var s = (ForallStmt)stmt;
-          s.IsGhost = mustBeErasable || s.Kind != ForallStmt.BodyKind.Assign || ExpressionTester.UsesSpecFeatures(s.Range);
-          if (proofContext != null && s.Kind == ForallStmt.BodyKind.Assign) {
-            Error(s, $"{proofContext} is not allowed to perform an aggregate heap update");
-          } else if (s.Body != null) {
-            Visit(s.Body, s.IsGhost, s.Kind == ForallStmt.BodyKind.Assign ? proofContext : "a forall statement");
-          }
-          s.IsGhost = s.IsGhost || s.Body == null || s.Body.IsGhost;
-
-          if (!s.IsGhost) {
-            // Since we've determined this is a non-ghost forall statement, we now check that the bound variables have compilable bounds.
-            var uncompilableBoundVars = s.UncompilableBoundVars();
-            if (uncompilableBoundVars.Count != 0) {
-              foreach (var bv in uncompilableBoundVars) {
-                Error(s, "forall statements in non-ghost contexts must be compilable, but Dafny's heuristics can't figure out how to produce or compile a bounded set of values for '{0}'", bv.Name);
-              }
-            }
-
-            // If there were features in the range that are treated differently in ghost and non-ghost
-            // contexts, make sure they get treated for non-ghost use.
-            ExpressionTester.CheckIsCompilable(resolver, s.Range, codeContext);
-          }
-
-        } else if (stmt is ModifyStmt) {
-          var s = (ModifyStmt)stmt;
-          if (proofContext != null) {
-            Error(stmt, $"a modify statement is not allowed in {proofContext}");
-          }
-
-          s.IsGhost = mustBeErasable;
-          if (s.IsGhost) {
-            s.Mod.Expressions.Iter(resolver.DisallowNonGhostFieldSpecifiers);
-          }
-          if (s.Body != null) {
-            Visit(s.Body, mustBeErasable, proofContext);
-          }
-
-        } else if (stmt is CalcStmt) {
-          var s = (CalcStmt)stmt;
-          s.IsGhost = true;
-          foreach (var h in s.Hints) {
-            Visit(h, true, "a hint");
-          }
-
-        } else if (stmt is MatchStmt) {
-          var s = (MatchStmt)stmt;
-          s.IsGhost = mustBeErasable || ExpressionTester.UsesSpecFeatures(s.Source) || ExpressionTester.FirstCaseThatDependsOnGhostCtor(s.Cases) != null;
-          if (!mustBeErasable && s.IsGhost) {
-            resolver.reporter.Info(MessageSource.Resolver, s.Tok, "ghost match");
-          }
-          s.Cases.Iter(kase => kase.Body.Iter(ss => Visit(ss, s.IsGhost, proofContext)));
-          s.IsGhost = s.IsGhost || s.Cases.All(kase => kase.Body.All(ss => ss.IsGhost));
-          if (!s.IsGhost) {
-            // If there were features in the source expression that are treated differently in ghost and non-ghost
-            // contexts, make sure they get treated for non-ghost use.
-            ExpressionTester.CheckIsCompilable(resolver, s.Source, codeContext);
-          }
-
-        } else if (stmt is ConcreteSyntaxStatement) {
-          var s = (ConcreteSyntaxStatement)stmt;
-          Visit(s.ResolvedStatement, mustBeErasable, proofContext);
-          s.IsGhost = s.IsGhost || s.ResolvedStatement.IsGhost;
-
-        } else if (stmt is SkeletonStatement) {
-          var s = (SkeletonStatement)stmt;
-          s.IsGhost = mustBeErasable;
-          if (s.S != null) {
-            Visit(s.S, mustBeErasable, proofContext);
-            s.IsGhost = s.IsGhost || s.S.IsGhost;
-          }
-
-        } else {
-          Contract.Assert(false); throw new cce.UnreachableException();
-        }
-      }
-
-      private void CheckAssignStmt(AssignStmt s, bool mustBeErasable, [CanBeNull] string proofContext) {
-        Contract.Requires(s != null);
-        Contract.Requires(mustBeErasable || proofContext == null);
-
-        var lhs = s.Lhs.Resolved;
-
-        // Make an auto-ghost variable a ghost if the RHS is a ghost
-        if (lhs.Resolved is AutoGhostIdentifierExpr autoGhostIdExpr) {
-          if (s.Rhs is ExprRhs eRhs && ExpressionTester.UsesSpecFeatures(eRhs.Expr)) {
-            autoGhostIdExpr.Var.MakeGhost();
-          } else if (s.Rhs is TypeRhs tRhs) {
-            if (tRhs.InitCall != null && tRhs.InitCall.Method.IsGhost) {
-              autoGhostIdExpr.Var.MakeGhost();
-            } else if (tRhs.ArrayDimensions != null && tRhs.ArrayDimensions.Exists(ExpressionTester.UsesSpecFeatures)) {
-              autoGhostIdExpr.Var.MakeGhost();
-            } else if (tRhs.ElementInit != null && ExpressionTester.UsesSpecFeatures(tRhs.ElementInit)) {
-              autoGhostIdExpr.Var.MakeGhost();
-            } else if (tRhs.InitDisplay != null && tRhs.InitDisplay.Any(ExpressionTester.UsesSpecFeatures)) {
-              autoGhostIdExpr.Var.MakeGhost();
-            }
-          }
-        }
-
-        if (proofContext != null && s.Rhs is TypeRhs) {
-          Error(s.Rhs.Tok, $"{proofContext} is not allowed to use 'new'");
-        }
-
-        var gk = AssignStmt.LhsIsToGhost_Which(lhs);
-        if (gk == AssignStmt.NonGhostKind.IsGhost) {
-          s.IsGhost = true;
-          if (proofContext != null && !(lhs is IdentifierExpr)) {
-            Error(lhs.tok, $"{proofContext} is not allowed to make heap updates");
-          }
-          if (s.Rhs is TypeRhs tRhs && tRhs.InitCall != null) {
-            Visit(tRhs.InitCall, true, proofContext);
-          }
-        } else if (gk == AssignStmt.NonGhostKind.Variable && codeContext.IsGhost) {
-          // cool
-        } else if (mustBeErasable) {
-          if (inConstructorInitializationPhase && codeContext is Constructor && codeContext.IsGhost && lhs is MemberSelectExpr mse &&
-              mse.Obj.Resolved is ThisExpr) {
-            // in this first division (before "new;") of a ghost constructor, allow assignment to non-ghost field of the object being constructed
-          } else {
-            string reason;
-            if (codeContext.IsGhost) {
-              reason = string.Format("this is a ghost {0}", codeContext is MemberDecl member ? member.WhatKind : "context");
-            } else {
-              reason = "the statement is in a ghost context; e.g., it may be guarded by a specification-only expression";
-            }
-            Error(s, $"assignment to {AssignStmt.NonGhostKind_To_String(gk)} is not allowed in this context, because {reason}");
-          }
-        } else {
-          if (gk == AssignStmt.NonGhostKind.Field) {
-            var mse = (MemberSelectExpr)lhs;
-            ExpressionTester.CheckIsCompilable(resolver, mse.Obj, codeContext);
-          } else if (gk == AssignStmt.NonGhostKind.ArrayElement) {
-            ExpressionTester.CheckIsCompilable(resolver, lhs, codeContext);
-          }
-
-          if (s.Rhs is ExprRhs) {
-            var rhs = (ExprRhs)s.Rhs;
-            if (!AssignStmt.LhsIsToGhost(lhs)) {
-              ExpressionTester.CheckIsCompilable(resolver, rhs.Expr, codeContext);
-            }
-          } else if (s.Rhs is HavocRhs) {
-            // cool
-          } else {
-            var rhs = (TypeRhs)s.Rhs;
-            if (rhs.ArrayDimensions != null) {
-              rhs.ArrayDimensions.ForEach(ee => ExpressionTester.CheckIsCompilable(resolver, ee, codeContext));
-              if (rhs.ElementInit != null) {
-                ExpressionTester.CheckIsCompilable(resolver, rhs.ElementInit, codeContext);
-              }
-              if (rhs.InitDisplay != null) {
-                rhs.InitDisplay.ForEach(ee => ExpressionTester.CheckIsCompilable(resolver, ee, codeContext));
-              }
-            }
-            if (rhs.InitCall != null) {
-              var callee = rhs.InitCall.Method;
-              if (callee.IsGhost) {
-                Error(rhs.InitCall, "the result of a ghost constructor can only be assigned to a ghost variable");
-              }
-              for (var i = 0; i < rhs.InitCall.Args.Count; i++) {
-                if (!callee.Ins[i].IsGhost) {
-                  ExpressionTester.CheckIsCompilable(resolver, rhs.InitCall.Args[i], codeContext);
-                }
-              }
-            }
-          }
-        }
-      }
-    }
     #endregion
 
     // ------------------------------------------------------------------------------------------------------
@@ -10369,7 +8663,7 @@ namespace Microsoft.Dafny {
     /// This method can be called even if the resolution of "fe" failed; in that case, this method will
     /// not issue any error message.
     /// </summary>
-    void DisallowNonGhostFieldSpecifiers(FrameExpression fe) {
+    public void DisallowNonGhostFieldSpecifiers(FrameExpression fe) {
       Contract.Requires(fe != null);
       if (fe.Field != null && !fe.Field.IsGhost) {
         reporter.Error(MessageSource.Resolver, fe.E, "in a ghost context, only ghost fields can be mentioned as modifies frame targets ({0})", fe.FieldName);
@@ -10841,32 +9135,6 @@ namespace Microsoft.Dafny {
         Opt = ResolveTypeOptionEnum.AllowPrefixExtend;
         Parent = parent;
       }
-    }
-
-    /// <summary>
-    /// If ResolveType/ResolveTypeLenient encounters a (datatype or class) type "C" with no supplied arguments, then
-    /// the ResolveTypeOption says what to do.  The last three options take a List as a parameter, which (would have
-    /// been supplied as an argument if C# had datatypes instead of just enums, but since C# doesn't) is supplied
-    /// as another parameter (called 'defaultTypeArguments') to ResolveType/ResolveTypeLenient.
-    /// </summary>
-    public enum ResolveTypeOptionEnum {
-      /// <summary>
-      /// never infer type arguments
-      /// </summary>
-      DontInfer,
-      /// <summary>
-      /// create a new InferredTypeProxy type for each needed argument
-      /// </summary>
-      InferTypeProxies,
-      /// <summary>
-      /// if at most defaultTypeArguments.Count type arguments are needed, use a prefix of defaultTypeArguments
-      /// </summary>
-      AllowPrefix,
-      /// <summary>
-      /// same as AllowPrefix, but if more than defaultTypeArguments.Count type arguments are needed, first
-      /// extend defaultTypeArguments to a sufficient length
-      /// </summary>
-      AllowPrefixExtend,
     }
 
     /// <summary>
@@ -14650,42 +12918,6 @@ namespace Microsoft.Dafny {
       return GetThisType(tok, (TopLevelDeclWithMembers)member.EnclosingClass);
     }
 
-    public record ResolutionContext(ICodeContext CodeContext, bool IsTwoState, bool InOld, bool InReveal,
-      bool InFunctionPostcondition, bool InFirstPhaseConstructor) {
-
-      // Invariants:
-      // InOld implies !IsTwoState
-      // InFirstPhaseConstructor implies codeContext is Constructor
-
-      public bool IsGhost => CodeContext.IsGhost;
-
-      public ResolutionContext(ICodeContext codeContext, bool isTwoState)
-        : this(codeContext, isTwoState, false, false, false, false) {
-      }
-
-      /// <summary>
-      /// Return a ResolutionContext appropriate for the body of "codeContext".
-      /// </summary>
-      public static ResolutionContext FromCodeContext(ICodeContext codeContext) {
-        bool isTwoState;
-        if (codeContext is NoContext || codeContext is DatatypeDecl) {
-          isTwoState = false;
-        } else if (codeContext is Function && !(codeContext is TwoStateFunction)) {
-          isTwoState = false;
-        } else {
-          isTwoState = true;
-        }
-        return new ResolutionContext(codeContext, isTwoState);
-      }
-
-      public ResolutionContext WithGhost(bool isGhost) {
-        if (CodeContext.IsGhost == isGhost) {
-          return this;
-        }
-        return new ResolutionContext(new CodeContextWrapper(CodeContext, isGhost), IsTwoState, InOld, InReveal, InFunctionPostcondition, InFirstPhaseConstructor);
-      }
-    }
-
     /// <summary>
     /// "IsTwoState" implies that "old" and "fresh" expressions are allowed.
     /// </summary>
@@ -18134,104 +16366,6 @@ namespace Microsoft.Dafny {
         }
       }
       return 0;
-    }
-  }
-
-  class Scope<Thing> where Thing : class {
-    [Rep]
-    readonly List<string> names = new List<string>();  // a null means a marker
-    [Rep]
-    readonly List<Thing> things = new List<Thing>();
-    [ContractInvariantMethod]
-    void ObjectInvariant() {
-      Contract.Invariant(names != null);
-      Contract.Invariant(things != null);
-      Contract.Invariant(names.Count == things.Count);
-      Contract.Invariant(-1 <= scopeSizeWhereInstancesWereDisallowed && scopeSizeWhereInstancesWereDisallowed <= names.Count);
-    }
-
-    int scopeSizeWhereInstancesWereDisallowed = -1;
-
-    public bool AllowInstance {
-      get { return scopeSizeWhereInstancesWereDisallowed == -1; }
-      set {
-        Contract.Requires(AllowInstance && !value);  // only allowed to change from true to false (that's all that's currently needed in Dafny); Pop is what can make the change in the other direction
-        scopeSizeWhereInstancesWereDisallowed = names.Count;
-      }
-    }
-
-    public void PushMarker() {
-      names.Add(null);
-      things.Add(null);
-    }
-
-    public void PopMarker() {
-      int n = names.Count;
-      while (true) {
-        n--;
-        if (names[n] == null) {
-          break;
-        }
-      }
-      names.RemoveRange(n, names.Count - n);
-      things.RemoveRange(n, things.Count - n);
-      if (names.Count < scopeSizeWhereInstancesWereDisallowed) {
-        scopeSizeWhereInstancesWereDisallowed = -1;
-      }
-    }
-
-    public enum PushResult { Duplicate, Shadow, Success }
-
-    /// <summary>
-    /// Pushes name-->thing association and returns "Success", if name has not already been pushed since the last marker.
-    /// If name already has been pushed since the last marker, does nothing and returns "Duplicate".
-    /// If the appropriate command-line option is supplied, then this method will also check if "name" shadows a previous
-    /// name; if it does, then it will return "Shadow" instead of "Success".
-    /// </summary>
-    public PushResult Push(string name, Thing thing) {
-      Contract.Requires(name != null);
-      Contract.Requires(thing != null);
-      if (Find(name, true) != null) {
-        return PushResult.Duplicate;
-      } else {
-        var r = PushResult.Success;
-        if (DafnyOptions.O.WarnShadowing && Find(name, false) != null) {
-          r = PushResult.Shadow;
-        }
-        names.Add(name);
-        things.Add(thing);
-        return r;
-      }
-    }
-
-    Thing Find(string name, bool topScopeOnly) {
-      Contract.Requires(name != null);
-      for (int n = names.Count; 0 <= --n;) {
-        if (names[n] == null) {
-          if (topScopeOnly) {
-            return null;  // not present
-          }
-        } else if (names[n] == name) {
-          Thing t = things[n];
-          Contract.Assert(t != null);
-          return t;
-        }
-      }
-      return null;  // not present
-    }
-
-    public Thing Find(string name) {
-      Contract.Requires(name != null);
-      return Find(name, false);
-    }
-
-    public Thing FindInCurrentScope(string name) {
-      Contract.Requires(name != null);
-      return Find(name, true);
-    }
-
-    public bool ContainsDecl(Thing t) {
-      return things.Exists(thing => thing == t);
     }
   }
